authorization implemented

+++ src/api/musics/handler.js
@@ -16,8 +16,9 @@ class MusicHandler {
     try {
       this._validator.validateMusicPayload(request.payload)
       const { title = 'untitled', year, performer, genre, duration } = request.payload
+      const { id: credentialId } = request.auth.credentials
 
-      const songId = await this._service.addMusic({ title, year, performer, genre, duration })
+      const songId = await this._service.addMusic({ title, year, performer, genre, duration, owner: credentialId })
 
       const response = h.response({
         status: 'success',
@@ -48,8 +49,10 @@ class MusicHandler {
     }
   }
 
-  async getAllMusicsHandler () {
-    const music = await this._service.getMusics()
+  async getAllMusicsHandler (request) {
+    const { id: credentialId } = request.auth.credentials
+
+    const music = await this._service.getMusics(credentialId)
     const song = music.map(m => ({
       id: m.id,
       title: m.title,
@@ -66,6 +69,9 @@ class MusicHandler {
   async getMusicByIdHandler (request, h) {
     try {
       const { songId } = request.params
+      const { id: credentialId } = request.auth.credentials
+
+      await this._service.verifyNoteOwner(songId, credentialId)
       const song = await this._service.getMusicById(songId)
 
       return {
@@ -101,7 +107,9 @@ class MusicHandler {
       this._validator.validateMusicPayload(request.payload)
       const { title, year, performer, genre, duration } = request.payload
       const { songId } = request.params
+      const { id: credentialId } = request.auth.credentials
 
+      await this._service.verifyMusicOwner(songId, credentialId)
       await this._service.editMusicById(songId, { title, year, performer, genre, duration })
 
       return {
@@ -131,6 +139,9 @@ class MusicHandler {
   async deleteMusicByIdHandler (request, h) {
     try {
       const { songId } = request.params
+      const { id: credentialId } = request.auth.credentials
+
+      await this._service.verifyNoteOwner(songId, credentialId)
       await this._service.deleteMusicById(songId)
 
       return {
